+++ be/src/vec/exec/scan/vscanner.h
@@ -124,7 +124,7 @@ public:
     int64_t get_scanner_wait_worker_timer() const { return _scanner_wait_worker_timer; }
 
     void update_scan_cpu_timer() {
-        int64_t cpu_time = _cpu_watch.elapsed_time();
+        int64_t cpu_time = (int64_t)_cpu_watch.elapsed_time();
         _scan_cpu_timer += cpu_time;
         _query_statistics->add_cpu_nanos(cpu_time);
     }
